+++ src/data/bramble.ts
@@ -1,4 +1,4 @@
-export const BUILD = '0.14'
+export const BUILD = '0.15'
 
 export const attributes = [
   { id: 'agility', name: 'Agility', description: 'Motion & Balance' },
@@ -25,7 +25,7 @@ export const homelands = [
   { name:'Wandering Road', skills:['Guile','Charm'] },
   { name:'Tidewake Harbor', skills:['Seeking','Tradeskill'] },
   { name:'Ancient Towers', skills:['Lorekeeping (Ancients)','Awareness'] },
-  { name:'Gloommere', skills:['Whisperstep','Herbalry'] },
+  { name:'Gloommere', skills:['Whisperster','Herbalry'] },
   { name:'Halerun Fields', skills:['Beastride','Stonegrasp'] },
   { name:'Deepmarket', skills:['Shadehand','Guile'] },
   { name:'Redmir Dunes', skills:['Wayfinding','Stonegrasp'] },